Removed old blockchain route
